[Warning][src/gfx/cockpit_xml.cpp][make]Expecting enum a, getting enum b; This was 'fixed' by casting 'elem' to (int) thus preventing the compiler from running the check. The code works as expected.

+++ engine/src/gfx/cockpit_xml.cpp
@@ -215,7 +215,7 @@ string getRes( string inp )
 }
 
 /*! beginElement
- * Sets up the entire cockpit. 
+ * Sets up the entire cockpit.
  */
 void GameCockpit::beginElement( const string &name, const AttributeList &attributes )
 {
@@ -246,7 +246,7 @@ void GameCockpit::beginElement( const string &name, const AttributeList &attribu
     VSSprite *oldpit       = NULL;
     bool      replaced[4]  = {false, false, false, false};
     int       counter      = 0;
-    switch (elem)
+    switch ((int)elem)
     {
     case CockpitXML::COCKPIT:
         for (iter = attributes.begin(); iter != attributes.end(); iter++) {
@@ -582,7 +582,7 @@ loadsprite:
 
 void GameCockpit::endElement( const string &name )
 {
-	//Nothing will be coming after the set up. 
+	//Nothing will be coming after the set up.
 }
 
 using namespace VSFileSystem;
